Proof of concept: ad-hoc entity types
This is primarily for #10753

This is a draft; comments, exceptions, some testing intentionally left not done yet.

The idea is to build and ad-hoc entity type which can then used for a raw SQL query. However, it actually also allows composing over the raw SQL query and ad-hoc LINQ queries.

Things to consider:

- The entity type cannot have relationships
- Properties are mapped by convention, but mapping attributes are respected. This is more important for LINQ queries than raw SQL queries since it allows EF to target the correct table and columns.
- The entity types are keyless. We could support keys, which would open up inserts, updates, and deletes, but I'm not sure we should.
- This happens automatically if the CLR type is not mapped. This means no new API service, but it also means it could be easy to accidentally use an ad-hoc type if you forget to add the type to the model. On the other hand, and type with navigations will fail anyway, so perhaps with a reasonable exception message its okay.